+++ app/services/projects/building_manual_builder.rb
@@ -10,17 +10,15 @@ module Projects
 
     private
 
-    def create_building
-      1.upto(buildings_count) do |_index|
-        project.site_address.each do |site_address|
-          project.buildings.build(
-            name: site_address['name'],
-            assignee: project.assignee,
-            apartments_count: site_address['apartments_count'],
-            move_in_starts_on: site_address['move_in_starts_on'],
-            address_attributes: site_address['address']
-          )
-        end
+    def create_building # rubocop:disable Metrics/AbcSize
+      1.upto(buildings_count) do |index|
+        project.buildings.build(
+          name: project.site_address[index - 1]['name'],
+          assignee: project.assignee,
+          apartments_count: project.site_address[index - 1]['apartments_count'],
+          move_in_starts_on: project.site_address[index - 1]['move_in_starts_on'],
+          address_attributes: project.site_address[index - 1]['address']
+        )
       end
     end
   end